created new cogs for admin, help and option commands
