+++ cogs/family.py
@@ -14,7 +14,8 @@ from interactions import User
 from .classes.custom_paginator import CustomPaginator
 from interactions.ext.paginators import Paginator
 from interactions import Client, ActionRow, Button, ButtonStyle, Intents, listen, slash_command, InteractionContext, SlashCommandOption, OptionType, slash_default_member_permission, SlashCommandChoice, context_menu, CommandType, Permissions, cooldown, Buckets, Embed, Extension
-
+import logging
+from logger import loggy
 
 emojis = ['👍', '👎']
 
@@ -24,7 +25,8 @@ class Family(Extension):
 
     @listen()
     async def on_ready(self):
-        print('Family Cog is ready!')
+        # print('Family Cog is ready!')
+        loggy.info('Family Cog is ready')
 
     async def cog_check(self, ctx):
         return await self.bot.validate_user(ctx)
