+++ cogs/guilds.py
@@ -10,6 +10,8 @@ from collections import ChainMap
 from interactions import User
 from interactions import Client, ActionRow, Button, ButtonStyle, Intents, listen, slash_command, InteractionContext, SlashCommandOption, OptionType, slash_default_member_permission, SlashCommandChoice, context_menu, CommandType, Permissions, cooldown, Buckets, Embed, Extension
 import uuid
+import logging
+from logger import loggy
 emojis = ['👍', '👎']
 
 class Teams(Extension):
@@ -18,7 +20,8 @@ class Teams(Extension):
 
     @listen()
     async def on_ready(self):
-        print('Teams Cog is ready!')
+        # print('Teams Cog is ready!')
+        loggy.info('Guild Cog is ready')
 
     async def cog_check(self, ctx):
         return await self.bot.validate_user(ctx)
