+++ cogs/lobbies.py
@@ -9,6 +9,8 @@ from PIL import Image, ImageFont, ImageDraw
 import requests
 from collections import ChainMap
 from interactions import Client, ActionRow, Button, ButtonStyle, Intents, listen, slash_command, InteractionContext, SlashCommandOption, OptionType, slash_default_member_permission, SlashCommandChoice, context_menu, CommandType, Permissions, cooldown, Buckets, Embed, Extension
+import logging
+from logger import loggy
 
 emojis = ['👍', '👎']
 
@@ -21,6 +23,7 @@ class Lobbies(Extension):
     @listen()
     async def on_ready(self):
         print('Lobbies Cog is ready!')
+        loggy.info('Lobbies Cog is ready')
 
     async def cog_check(self, ctx):
         return await self.bot.validate_user(ctx)
