+++ cogs/quests.py
@@ -6,6 +6,8 @@ import messages as m
 from interactions import User
 import asyncio
 from interactions import Client, ActionRow, Button, ButtonStyle, Intents, listen, slash_command, InteractionContext, SlashCommandOption, OptionType, slash_default_member_permission, SlashCommandChoice, context_menu, CommandType, Permissions, cooldown, Buckets, Embed, Extension
+import logging
+from logger import loggy
 
 class Quests(Extension):
     def __init__(self, bot):
@@ -14,6 +16,7 @@ class Quests(Extension):
     @listen()
     async def on_ready(self):
         print('Quests Cog is ready!')
+        loggy.info('Quests Cog is ready!')
 
     async def cog_check(self, ctx):
         return await self.bot.validate_user(ctx)
