+++ cogs/reward_drops.py
@@ -13,6 +13,8 @@ import random
 from logger import loggy
 from .quests import Quests
 from interactions import Client, ActionRow, Button, ButtonStyle, Intents, listen, slash_command, InteractionContext, SlashCommandOption, OptionType, slash_default_member_permission, SlashCommandChoice, context_menu, CommandType, Permissions, cooldown, Buckets, Embed, Extension
+import logging
+from logger import loggy
 
 class RewardDrops(Extension):
     def __init__(self, bot):
@@ -20,7 +22,8 @@ class RewardDrops(Extension):
 
     @listen()
     async def on_ready(self):
-        print('Drops Cog is ready!')
+        # print('Drops Cog is ready!')
+        loggy.info('Drops Cog is ready')
 
 
     async def cog_check(self, ctx):
